Kernel/aarch64: Move hardware clock configuration to own file
Separating the hardware interrupt stuff from the hardware clock
configuration allows us to use these utilities from the prekernel.
