+++ Kernel/CMakeLists.txt
@@ -444,6 +444,7 @@ elseif("${SERENITY_ARCH}" STREQUAL "aarch64")
         Arch/aarch64/RPi/Mailbox.cpp
         Arch/aarch64/RPi/MiniUART.cpp
         Arch/aarch64/RPi/SDHostController.cpp
+        Arch/aarch64/RPi/HardwareClocks.cpp
         Arch/aarch64/RPi/Timer.cpp
         Arch/aarch64/RPi/UART.cpp
         Arch/aarch64/RPi/Watchdog.cpp
